# perf: lazy calculate asset.source