<!--
  Thank you for submitting a pull request!

  We appreciate the time and effort you have invested in making these changes. Please ensure that you provide enough information to allow others to review your pull request.

  Upon submission, your pull request will be automatically assigned with reviewers.

  If you want to learn more about contributing to this project, please visit: https://github.com/web-infra-dev/rspack/blob/main/CONTRIBUTING.md.
-->

## Summary

This PR will add lazy calculation for asset.source which will make html-webpack-plugin faster.

Here is some data from my project.

| html-webpack-plugin | pr + html-webpack-plugin | rspack.HtmlRspackPlugin |
|--|--|--|
| 1s 636ms | 640ms | 580ms |
| ![image](https://github.com/web-infra-dev/rspack/assets/9291503/fddb36d2-b5bc-49d2-8153-7a7037096058) | ![image](https://github.com/web-infra-dev/rspack/assets/9291503/91f3b433-33db-436f-884c-fe9a0206da43) | ![image](https://github.com/web-infra-dev/rspack/assets/9291503/673388c8-4edf-48fe-a03b-d7ebd7863591) |



<!-- Can you explain the reasoning behind implementing this change? What problem or issue does this pull request resolve? -->

<!-- It would be helpful if you could provide any relevant context, such as GitHub issues or related discussions. -->

## Test Plan

<!-- Can you please describe how you tested the changes you made to the code? -->

## Require Documentation?

<!-- Does this PR require documentation? -->

- [ ] No
- [ ] Yes, the corresponding rspack-website PR is \_\_